move contentGraphNode and Edge in their own package
