+++ src/main/java/org/sbml4j/model/sbml/ext/sbml4j/BiomodelsQualifier.java
@@ -17,7 +17,7 @@ import org.neo4j.ogm.annotation.EndNode;
 import org.neo4j.ogm.annotation.RelationshipEntity;
 import org.neo4j.ogm.annotation.StartNode;
 import org.sbml.jsbml.CVTerm;
-import org.sbml4j.model.common.ContentGraphEdge;
+import org.sbml4j.model.contentgraph.ContentGraphEdge;
 import org.sbml4j.model.sbml.SBMLSBaseEntity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
